+++ src/fields/OtpField/index.ts
@@ -1,2 +1,3 @@
 export * from './OtpField'
 export { otpFieldClasses } from './OtpFieldControl'
+export * from './ResendCodeButton'